chore(ci): add a simple and faster CI
This CI is better suited for projects with fewer lines of code.
There is no need to parallelize the CI until a sufficient amount of code
is present in the project.

The recommended path is to first use the simple CI and then switch to
the complex CI when the project grows.