chore(ci): make sure examples are always checked in the merge queue
The only way to make sure that the check examples is a required status check
for the merge queue is to add it as well in the PR branch protection required status checks.

These changes will make sure that the check examples is always run, even if the PR does not
changes any of the files that trigger the check examples job.

Moreover, the check examples workflow will always run in the merge queue, as
changes in all the packages can make the examples fail.
See https://github.com/swarmion/swarmion/actions/runs/4651813666 for an example.