chore(sonarcloud): greatly simplify both sonar and ci
Now we have a single sonarcloud project, which is the monorepo itself.
This allows us to greatly simplify the CI by having a single job that relies
completely on the Nx cache.
Moreover, the sonarcloud scan will now run on every PR and thus its status
check can be enforced going further.